+++ src/core/field/migration.rs
@@ -8,4 +8,5 @@ pub(crate) struct FieldMigration {
     pub(crate) default: Option<Value>,
     pub(crate) action: Option<Pipeline>,
     pub(crate) priority: Option<usize>,
+    pub(crate) drop: bool,
 }