+++ src/parser/std/decorators/field/migration_decorator.rs
@@ -3,10 +3,10 @@ use crate::core::field::migration::FieldMigration;
 use crate::parser::ast::argument::Argument;
 use crate::prelude::Value;
 
-static VALID_NAMES: [&str; 5] = ["renamed", "default", "version", "action", "priority"];
+static VALID_NAMES: [&str; 6] = ["renamed", "default", "version", "action", "priority", "drop"];
 
 pub(crate) fn migration_decorator(args: Vec<Argument>, field: &mut Field) {
-    let mut migration = FieldMigration { renamed: vec![], default: None, version: None, action: None, priority: None };
+    let mut migration = FieldMigration { renamed: vec![], default: None, version: None, action: None, priority: None, drop: false };
     for arg in args {
         if arg.name.is_none() {
             panic!("@migration requires argument name.");
@@ -45,6 +45,11 @@ pub(crate) fn migration_decorator(args: Vec<Argument>, field: &mut Field) {
                 let value = arg.resolved.as_ref().unwrap().as_value().unwrap();
                 let u = value.as_usize().unwrap();
                 migration.priority = Some(u);
+            }
+            "drop" => {
+                let value = arg.resolved.as_ref().unwrap().as_value().unwrap();
+                let b = value.as_bool().unwrap();
+                migration.drop = b;
             }
             _ => unreachable!()
         }